refactor: controller의 dto service로 이동
서비스에서 Response 객체를 생성해서 컨트롤러에게 전달해주기 때문에,
dto 패키지를 controller에서 service 패키지로 옮겼다.

그리고 GenderWordFrequencyResponse와 WordFrequencyResponse가 같은
WordFrequency를 쓰기 때문에,
WordFrequencyResponse에 있는 WordFrequency 클래스를 별도의 클래스로
분리해줬다.
