+++ src/main/java/com/example/spinlog/statistics/controller/MBTIStatisticsController.java
@@ -1,9 +1,9 @@
 package com.example.spinlog.statistics.controller;
 
 import com.example.spinlog.article.entity.RegisterType;
-import com.example.spinlog.statistics.controller.dto.MBTIDailyAmountSumResponse;
-import com.example.spinlog.statistics.controller.dto.MBTIEmotionAmountAverageResponse;
-import com.example.spinlog.statistics.controller.dto.WordFrequencyResponse;
+import com.example.spinlog.statistics.service.dto.MBTIDailyAmountSumResponse;
+import com.example.spinlog.statistics.service.dto.MBTIEmotionAmountAverageResponse;
+import com.example.spinlog.statistics.service.dto.WordFrequencyResponse;
 import com.example.spinlog.statistics.repository.dto.MBTISatisfactionAverageDto;
 import com.example.spinlog.statistics.service.MBTIStatisticsService;
 import lombok.RequiredArgsConstructor;
