+++ src/main/java/com/example/spinlog/statistics/service/MBTIStatisticsService.java
@@ -1,9 +1,9 @@
 package com.example.spinlog.statistics.service;
 
 import com.example.spinlog.article.entity.RegisterType;
-import com.example.spinlog.statistics.controller.dto.MBTIDailyAmountSumResponse;
-import com.example.spinlog.statistics.controller.dto.MBTIEmotionAmountAverageResponse;
-import com.example.spinlog.statistics.controller.dto.WordFrequencyResponse;
+import com.example.spinlog.statistics.service.dto.MBTIDailyAmountSumResponse;
+import com.example.spinlog.statistics.service.dto.MBTIEmotionAmountAverageResponse;
+import com.example.spinlog.statistics.service.dto.WordFrequencyResponse;
 import com.example.spinlog.statistics.repository.MBTIStatisticsRepository;
 import com.example.spinlog.statistics.repository.dto.MBTIDailyAmountSumDto;
 import com.example.spinlog.statistics.repository.dto.MBTIEmotionAmountAverageDto;
@@ -17,7 +17,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.groupingBy;
