+++ src/main/java/com/example/spinlog/statistics/service/dto/WordFrequency.java
@@ -1,0 +1,12 @@
+package com.example.spinlog.statistics.service.dto;
+
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class WordFrequency{
+    private String word;
+    private Long frequency;
+}
